Design: SnsShare 공통 컴포넌트 생성
