+++ src/app/(hasNotMap)/youth-place/_components/PlaceCard/PlaceCard.module.scss
@@ -66,18 +66,7 @@
   color: $text-secondary-color;
   font-size: 14px;
 }
-.shareIcon {
-  position: absolute;
-  top: 0;
-  right: 0;
-  &:hover {
-    .icon {
-      path {
-        fill: $icon-interactive-primary-color; // todo
-      }
-    }
-  }
-}
+
 .cost {
   margin-top: 19px;
   font-size: 14px; // todo
