+++ src/app/(hasNotMap)/youth-policy/_components/PolicyGridCard/index.tsx
@@ -1,14 +1,17 @@
+'use client';
+
 import styles from './PolicyGridCard.module.scss';
+import { useRouter } from 'next/navigation';
 import { SimplePolicy } from '@/types/policy';
-import Link from 'next/link';
-import ShareSvg from '/public/icons/share-2.svg';
 import Image from 'next/image';
+import SnsShare from '@/components/policy_place/SnsShare';
 
 type Props = {
   policy: SimplePolicy;
 };
 
 export default function PolicyGridCard({ policy }: Props) {
+  const router = useRouter();
   const {
     policyId,
     title,
@@ -20,12 +23,9 @@ export default function PolicyGridCard({ policy }: Props) {
     recommendCount,
   } = policy;
 
-  // TODO
-  const onShare = () => console.log('공유 버튼 클릭! - 오픈 그래프 카드');
-
   return (
-    <Link href={`/youth-policy/${policyId}`}>
-      <section className={styles.container}>
+    <section className={styles.container} onClick={() => router.push(`/youth-policy/${policyId}`)}>
+      <div className={styles.infoWrapper}>
         <div className={styles.iconWrapper}>
           <div className={styles.icon}>
             <Image src="/icons/eye.svg" alt="추천 아이콘" width={12} height={12} />
@@ -36,19 +36,17 @@ export default function PolicyGridCard({ policy }: Props) {
             <span>{recommendCount}</span>
           </div>
         </div>
-        <button className={styles.shareButton} onClick={onShare}>
-          <ShareSvg className={styles.icon} width="16" height="16" />
-        </button>
-        <div className={styles.contents}>
-          <p className={styles.title}>{title}</p>
-          <p className={styles.description}>{description}</p>
-          <p className={styles.organization}>{organization}</p>
-        </div>
-        <div className={styles.bottomArea}>
-          <p className={styles.progress}>{progress}</p>
-          <p className={styles.field}>{policyField}</p>
-        </div>
-      </section>
-    </Link>
+        <SnsShare />
+      </div>
+      <div className={styles.contents}>
+        <p className={styles.title}>{title}</p>
+        <p className={styles.description}>{description}</p>
+        <p className={styles.organization}>{organization}</p>
+      </div>
+      <div className={styles.bottomArea}>
+        <p className={styles.progress}>{progress}</p>
+        <p className={styles.field}>{policyField}</p>
+      </div>
+    </section>
   );
 }
