+++ src/components/policy_place/SnsShare/SnsShare.module.scss
@@ -1,0 +1,13 @@
+.shareIcon {
+  position: absolute;
+  top: 0;
+  right: 0;
+  padding: 0;
+  &:hover {
+    .icon {
+      path {
+        fill: $icon-interactive-primary-color;
+      }
+    }
+  }
+}
